requirements.txt
contourpy==1.0.6
cycler==0.11.0
easygui==0.98.3
fonttools==4.38.0
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
